+++ src/CashFlow.Domain/Repositories/Expenses/IExpensesReadOnlyRepository.cs
@@ -6,5 +6,5 @@ public interface IExpensesReadOnlyRepository
 {
     Task<List<Expense>> GetAll();
     
-    Task<Expense?> GetById(long id);
+    Task<Expense?> GetById(int id);
 }
